% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcs.R
\name{rcs}
\alias{rcs}
\alias{drcs}
\title{Calculate basis terms for restricted cubic splines}
\usage{
rcs(x, df = 3, iknots = NULL, bknots = NULL)

drcs(x, df = 3, iknots, bknots = NULL)
}
\arguments{
\item{x}{A numeric vector for which to evaluate the spline basis terms.}

\item{df}{Integer specifying the degrees of freedom for the splines. If
the \code{iknots} argument is \code{NULL}, then \code{df} - 1
internal knots are placed at equally spaced percentiles of \code{x}.}

\item{iknots}{An optional vector of internal knot positions. Note that the
\code{df} argument is ignored if \code{iknots} is specified.}

\item{bknots}{A numeric vector of length 2 specifying the position
of the lower and upper boundary knots, beyond which a linearity
restriction is enforced. If \code{NULL}, then the boundary knots are set
to the minimum and maximum of \code{x}.}
}
\value{
A matrix, with dimension \code{length(x)} by \code{df}, where
  \code{df} is one greater than the number of internal knots.
}
\description{
The \code{rcs} function calculates the basis terms for the restricted
cubic splines as defined in Royston and Parmar (2002). The \code{drcs}
function calculates the corresponding first derivative of the basis terms.
}
\examples{
times <- runif(40,0,10)

rcs1 <- rcs(times, df = 2) # one internal knot at median
rcs2 <- rcs(times, iknots = c(2,5,8))
rcs3 <- rcs(times, iknots = c(2,5), bknots = c(1,8))

drcs1 <- drcs(times, df = 2)
drcs2 <- drcs(times, iknots = c(2,5,8))

}
\references{
Royston P, Parmar MKB. Flexible parametric proportional-hazards
  and proportional-odds models for censored survival data, with application
  to prognostic modelling and estimation of treatment effects.
  \emph{Statistics in Medicine}. 2002;21:2175--2197. \doi{10.1002/sim.1203}
}
